Moved code for selecting words to separate module, fixasr-utilities. Added lastedit property to fixasr data.
