+++ src/Client/BrowserApp/src/client/app/fixasr/fixasr.component.ts
@@ -4,10 +4,11 @@ import { AsrSegment } from './asrsegment';
 import { AsrText } from './asrtext';
 import { AsrService } from './asr.service';
 import { VideoComponent } from '../video/video.component';
+import { FixasrUtilities } from './fixasr-utilities'
 
 // test
-import { ElementRef, Renderer } from '@angular/core';
-import { HostListener } from '@angular/core';
+import { ElementRef} from '@angular/core';
+//import { HostListener } from '@angular/core';
 
 //2017-02-18
 // https://angular.io/docs/ts/latest/guide/user-input.html
@@ -22,44 +23,62 @@ import { HostListener } from '@angular/core';
   templateUrl: './fixasr.component.html',
   styleUrls: ['./fixasr.component.css'],
     providers: [
-        AsrService
+        AsrService,
+        FixasrUtilities
     ]
 })
-export class FixasrComponent  implements OnInit, AfterViewInit {
-    //isTest : boolean = true;
+export class FixasrComponent  implements OnInit {
     errorMessage: string;
-    title = 'fixasr works!';
-    asr : AsrSegment[];
+    asrtext : AsrText;
+    asr: AsrSegment[];
     lastPhrasePlayed : number = 0;
     currentIndex : number = -1;
     isTyping : boolean = false;
     firstSpace : boolean = false;
     isInsertMode: boolean = false;
     modeButtonText: string = "REPLACE"
+    _scrollList: HTMLElement;
+    scrollPosition: number;
 
     // https://github.com/videogular/videogular2/blob/master/docs/using-the-api.md
 
     @ViewChild(VideoComponent)
     private videoComponent : VideoComponent;
 
-    // test
-    @ViewChild('myInput') input: ElementRef;
-    _el : HTMLElement;
-    _scrollList: HTMLElement;
+   @ViewChild('myInput') input: ElementRef;
 
-    // test
-    // constructor(private _asrService: AsrService) {
-    constructor(private _asrService: AsrService, private renderer: Renderer) {
+    constructor(
+        private _asrService: AsrService,
+        private _Utilities: FixasrUtilities) {
+    }
+
+    ngOnInit() {
+        this.getAsr();
         this._scrollList = document.getElementById('scroll-text');
     }
 
+    getpos() {
+        this.scrollPosition = this.getScrollPosition();
+    }
+    
+    setpos() {
+        this._scrollList.scrollTop = this.scrollPosition;
+    }
+    
     setScrollPosition(top: number){
-            this._scrollList.scrollTop = top;
+        this._scrollList.scrollTop = top;
     }
 
-    getScrollPosition(top: number) : number {
-            return this._scrollList.scrollTop;
+    getScrollPosition() : number {
+        return this._scrollList.scrollTop;
     }
+
+    toggleInsertMode(){
+        this.isInsertMode = !this.isInsertMode;
+        this.modeButtonText = (this.isInsertMode ? "INSERT" : "REPLACE");
+    }
+
+// #########################  Event Handlers ################################################
 
     onFocus(event: any, i : number){
         console.log("onFocus index=" + i + "  size=" + this.asr.length);
@@ -68,80 +87,16 @@ export class FixasrComponent  implements OnInit, AfterViewInit {
         this.firstSpace = false;
     }
 
-    toggleInsertMode(){
-        this.isInsertMode = !this.isInsertMode;
-        this.modeButtonText = (this.isInsertMode ? "INSERT" : "REPLACE");
-    }
-
-    // test
-    doMouseup(event: any, i : number) {
+    onMouseup(event: any, i : number) {
         if (this.isInsertMode) return;
 
-        // console.log(window.getSelection());
         var ele : HTMLInputElement = (<HTMLInputElement>event.target);
-        console.log("doMouseup index=" + i + "   start=" + ele.selectionStart + "   end=" + ele.selectionEnd);
-        console.log("doMouseup value=" + ele.value);
-        // ele.setSelectionRange(4,8);
-        this.selectWord(ele);
-        // if index is new, play phrase
-        // select word clicked on, or next word if not on word
+        console.log("onMouseup index=" + i + "   start=" + ele.selectionStart + "   end=" + ele.selectionEnd);
+        console.log("onMouseup value=" + ele.value);
+        this._Utilities.selectWord(ele);
     }
 
-    gotoNextInputElement(ele : HTMLInputElement, i : number) {
-        // if on last element, select last word of this element
-        if (i >= (this.asr.length - 1)) {
-            this.selectLastWord(ele);
-            return;
-        }
-
-        // ele is this <input> element. ele.parentElement is the <div> enclosing just this <input>.
-        // That <div>'s nextSibling is the <div> whose sole child is the next <input>.
-        var next : HTMLInputElement = <HTMLInputElement>(ele.parentElement.nextSibling.children[0]);
-        console.log(next.value);
-
-        if (typeof next.setSelectionRange != 'undefined') {
-            this.selectFirstWord(next);
-        }
-    }
-
-    gotoPriorInputElement(ele : HTMLInputElement, i : number) {
-        // if on first element, just return
-        if (i == 0) {
-            this.selectFirstWord(ele);
-            return;
-        }
-        var prior : HTMLInputElement = <HTMLInputElement>(ele.parentElement.previousSibling.children[0]);
-        if (typeof prior.setSelectionRange != 'undefined') {
-            this.selectFirstWord(prior);
-        }
-    }
-
-    gotoLastWordInPriorInputElement(ele : HTMLInputElement, i : number) {
-        // if on first element, select first word in this element
-        if (i == 0) {
-            this.selectFirstWord(ele);
-            return;
-        }
-
-        var prior : HTMLInputElement = <HTMLInputElement>(ele.parentElement.previousSibling.children[0]);
-        if (typeof prior.setSelectionRange != 'undefined') {
-            this.selectLastWord(prior);
-        }
-    }
-    gotoNextWord(ele : HTMLInputElement) {
-        var end : number = ele.selectionEnd;
-        ele.setSelectionRange(end + 2, end + 2);
-        this.selectWord(ele);
-    }
-
-    gotoPriorWord(ele : HTMLInputElement) {
-        var start : number = ele.selectionStart;
-        ele.setSelectionRange(start - 2, start - 2);
-        this.selectWord(ele);
-    }
-
-
-    doKey(event: any, i : number) {
+    onKey(event: any, i : number) {
         if (event.key == 'Insert') {
             this.toggleInsertMode();
             return;
@@ -159,32 +114,38 @@ export class FixasrComponent  implements OnInit, AfterViewInit {
         console.log("1 doKey index=" + i + "   start=" + start + "   end=" + end);
 
         switch (key) {
+
+        // If "Enter" play that portion of video
             case 'Enter':
                 this.playPhrase(i);
                 return;
+
+        // Handle the arrow keys
             case 'ArrowRight':
                 if (end  < value.length) {
-                    this.gotoNextWord(ele);
+                    this._Utilities.gotoNextWord(ele);
                 } else {
-                    this.gotoNextInputElement(ele, i);
+                    this._Utilities.gotoNextInputElement(ele, i);
                 }
                 return;
             case 'ArrowLeft':
                 if (start != 0) {
-                    this.gotoPriorWord(ele);
+                    this._Utilities.gotoPriorWord(ele);
                 } else {
-                    this.gotoLastWordInPriorInputElement(ele, i);
+                    this._Utilities.gotoLastWordInPriorInputElement(ele, i);
                 }
                 return;
             case 'ArrowDown':
-                this.gotoNextInputElement(ele, i);
+                if (i >= (this.asr.length - 1)) {
+                    this._Utilities.selectLastWord(ele);
+                } else {
+                    this._Utilities.gotoNextInputElement(ele, i);
+                }
                 return;
             case 'ArrowUp':
-                this.gotoPriorInputElement(ele, i);
+                this._Utilities.gotoPriorInputElement(ele, i);
                 return;
         }
-
-
 
         // If not space, set isTyping true and firstSpace false and return
         if (key != " ") {
@@ -228,6 +189,8 @@ export class FixasrComponent  implements OnInit, AfterViewInit {
             console.log("3 doKey start=" + start + "   value='" + value + "'")
         }
 
+        // Select the correct word
+        
         // "start" now is just before the next word, or at end of text.
         if (start == value.length) {
             if (value.charAt(start - 1) === " ") {
@@ -236,28 +199,34 @@ export class FixasrComponent  implements OnInit, AfterViewInit {
             }
         }
         ele.value = value;
-
-
         ele.setSelectionRange(start+2, start+2)
         console.log("start=" + ele.selectionStart + "   end=" + ele.selectionEnd);
-        this.selectWord(ele);
+        this._Utilities.selectWord(ele);
 
     }
 
-    ngOnInit() {
-        this.getAsr();
-        //this.getAsr(this.isTest);
+// #####################  get & put data via AsrService ##########################
 
+    getAsr() {
+        this._asrService.getAsr()
+            .subscribe(
+            asrtext => {
+                this.asr = asrtext.asrsegments;
+                this.asrtext.lastedit = asrtext.lastedit;
+            },
+            error => this.errorMessage = <any>error);
     }
 
-    ngAfterViewInit() {
-        //console.log('In ngAfterViewInit');
-
-        // test
-        //console.log(this.input);
-        //this._el = this.input.nativeElement;
-        //this.renderer.invokeElementMethod(this._el, 'focus');
+    saveChanges() {
+        console.log('saveTranscript');
+        this._asrService.postChanges(this.asrtext)
+            .subscribe (
+                t => t
+            );
+        //error => this.errorMessage = <any>error);
     }
+
+// #####################  Play selected part of video ####################################
 
     playPhrase(i : number) {
         if (i == this.lastPhrasePlayed) return;
@@ -284,100 +253,6 @@ export class FixasrComponent  implements OnInit, AfterViewInit {
         }
         this.videoComponent.playPhrase(start, duration);
     }
-
-    getAsr() {
-        this._asrService.getAsr()
-            .subscribe(
-            talks => this.asr = talks,
-            error => this.errorMessage = <any>error);
-    }
-
-    //getAsr(isTest: boolean) {
-    //    if (isTest) {
-    //        this.asr = this._asrService.getAsrFromMemory().data;
-    //        console.log('return from getAsrFromMemory');
-    //    } else {
-    //        this._asrService.getAsr()
-    //        .subscribe(
-    //        talks => this.asr = talks,
-    //        error => this.errorMessage = <any>error);
-    //    }
-    //}
-
-    saveChanges() {
-        console.log('saveTranscript');
-        //this._talkService.postChanges(this.talks)
-        this._asrService.postChanges(this.asr)
-            .subscribe (
-                t => t
-            );
-        //error => this.errorMessage = <any>error);
-    }
-
-    private selectFirstWord(ele: HTMLInputElement) {
-        ele.setSelectionRange(1, 1);
-        this.selectWord(ele);
-        ele.focus();
-    }
-
-    private selectLastWord(ele: HTMLInputElement) {
-        var len = ele.value.length
-        ele.setSelectionRange(len - 1, len - 1);
-        this.selectWord(ele);
-        ele.focus();
-    }
-
-    private selectWord(ele: HTMLInputElement) {
-        var start = ele.selectionStart;
-        var end = ele.selectionEnd;
-        var content = ele.value;
-        var contentLen = content.length;
-        console.log('selectWord start=' + start + '   end=' + end);
-
-        // If they actually selected some text, ignore
-        // This also ignores a prior select that we set here. This has 
-        // the effect of toggling the select on a word by clicking
-        // on it multiple times.
-        if (start != end) {
-             console.log('selectWord ignore if current select');
-           return;
-        } 
-
-        // find start of word
-        while (start > 0 && this.isNotSpace(content.charAt(start -1))) {
-          start--;
-        }
-        // find end of word
-        while (end < contentLen && this.isNotSpace(content.charAt(end))) {
-          end++;
-        }
-
-        // Unless it's a one-letter word, if they clicked at start or end of text, ignore.
-        //if (start + 1 != end) {
-        //    if ((start == 0) || (start == contentLen) ||
-        //        (content.charAt(start) == " ") || (content.charAt(start - 1) == " ")) {
-        //        console.log('selectWord ignore if start or end of word');
-        //        return;
-        //    }
-        //}
-
-        ele.setSelectionRange(start, end);
-        console.log('selectWord newstart=' + start + '   newend=' + end);
-   }
-
-   private isNotSpace(ch: string) {
-        return (ch != ' ');
-   }
-
-   private isAlphaNum(ch: string) {
-       var code = ch.charCodeAt(0);
-        if (!(code > 47 && code < 58) && // numeric (0-9)
-            !(code > 64 && code < 91) && // upper alpha (A-Z)
-            !(code > 96 && code < 123)) { // lower alpha (a-z)
-        return false;
-        }
-        return true;
-   }
 
    convertToSeconds(time: string) {
        var array = time.split(':');
